+++ controller/hybridgateway/watch/watch.go
@@ -79,6 +79,10 @@ func Watches(obj client.Object, cl client.Client) []Watcher {
 				&corev1.Secret{},
 			},
 			{
+				MapHTTPRouteForPluginConfigSecret(cl),
+				&corev1.Secret{},
+			},
+			{
 				MapRouteForKongResource[*configurationv1alpha1.KongCertificate](kindHTTPRoute),
 				&configurationv1alpha1.KongCertificate{},
 			},
@@ -139,6 +143,10 @@ func Watches(obj client.Object, cl client.Client) []Watcher {
 			},
 			{
 				MapGRPCRouteForClientCertSecret(cl),
+				&corev1.Secret{},
+			},
+			{
+				MapGRPCRouteForPluginConfigSecret(cl),
 				&corev1.Secret{},
 			},
 			{