+++ controller/hybridgateway/watch/watch_test.go
@@ -26,7 +26,7 @@ func TestWatches(t *testing.T) {
 		{
 			name:    "HTTPRoute with ReferenceGrant enabled",
 			obj:     &gwtypes.HTTPRoute{},
-			wantLen: 15,
+			wantLen: 16,
 			wantType: []any{
 				&gwtypes.Gateway{},
 				&gwtypes.GatewayClass{},
@@ -41,6 +41,7 @@ func TestWatches(t *testing.T) {
 				&gwtypes.ReferenceGrant{},
 				&configurationv1beta1.KongUpstreamPolicy{},
 				&corev1.Secret{},
+				&corev1.Secret{},
 				&configurationv1alpha1.KongCertificate{},
 				&configurationv1alpha1.KongReferenceGrant{},
 			},
@@ -48,7 +49,7 @@ func TestWatches(t *testing.T) {
 		{
 			name:    "GRPCRoute",
 			obj:     &gwtypes.GRPCRoute{},
-			wantLen: 15,
+			wantLen: 16,
 			wantType: []any{
 				&gwtypes.Gateway{},
 				&gwtypes.GatewayClass{},
@@ -62,6 +63,7 @@ func TestWatches(t *testing.T) {
 				&configurationv1alpha1.KongPluginBinding{},
 				&gwtypes.ReferenceGrant{},
 				&configurationv1beta1.KongUpstreamPolicy{},
+				&corev1.Secret{},
 				&corev1.Secret{},
 				&configurationv1alpha1.KongCertificate{},
 				&configurationv1alpha1.KongReferenceGrant{},