+++ controller/konnect/ops/sdk/zz_generated_sdkfactory.go
@@ -10,6 +10,8 @@ import (
 type GeneratedSDK interface {
 	GetAIGatewayAgentsSDK() sdkkonnectgo.AIGatewayAgentsSDK
 	GetAIGatewayAuthStrategiesSDK() sdkkonnectgo.AIGatewayAuthStrategiesSDK
+	GetAIGatewayCACertificatesSDK() sdkkonnectgo.AIGatewayCACertificatesSDK
+	GetAIGatewayCertificatesSDK() sdkkonnectgo.AIGatewayCertificatesSDK
 	GetAIGatewayConsumersSDK() sdkkonnectgo.AIGatewayConsumersSDK
 	GetAIGatewayConsumerGroupsSDK() sdkkonnectgo.AIGatewayConsumerGroupsSDK
 	GetAIGatewayDataPlaneCertificatesSDK() sdkkonnectgo.AIGatewayDataPlaneCertificatesSDK
@@ -47,6 +49,16 @@ func (w sdkWrapper) GetAIGatewayAgentsSDK() sdkkonnectgo.AIGatewayAgentsSDK {
 // GetAIGatewayAuthStrategiesSDK returns the SDK to operate AIGatewayAuthStrategy.
 func (w sdkWrapper) GetAIGatewayAuthStrategiesSDK() sdkkonnectgo.AIGatewayAuthStrategiesSDK {
 	return w.sdk.AIGatewayAuthStrategies
+}
+
+// GetAIGatewayCACertificatesSDK returns the SDK to operate AIGatewayCACertificate.
+func (w sdkWrapper) GetAIGatewayCACertificatesSDK() sdkkonnectgo.AIGatewayCACertificatesSDK {
+	return w.sdk.AIGatewayCACertificates
+}
+
+// GetAIGatewayCertificatesSDK returns the SDK to operate AIGatewayCertificate.
+func (w sdkWrapper) GetAIGatewayCertificatesSDK() sdkkonnectgo.AIGatewayCertificatesSDK {
+	return w.sdk.AIGatewayCertificates
 }
 
 // GetAIGatewayConsumersSDK returns the SDK to operate AIGatewayConsumer.