+++ controller/konnect/zz_generated_watch.go
@@ -26,6 +26,10 @@ func reconciliationWatchOptionsForEntity[
 		return AIGatewayAgentReconciliationWatchOptions(cl)
 	case *aiconfigurationv1alpha1.AIGatewayAuthStrategy:
 		return AIGatewayAuthStrategyReconciliationWatchOptions(cl)
+	case *aiconfigurationv1alpha1.AIGatewayCACertificate:
+		return AIGatewayCACertificateReconciliationWatchOptions(cl)
+	case *aiconfigurationv1alpha1.AIGatewayCertificate:
+		return AIGatewayCertificateReconciliationWatchOptions(cl)
 	case *aiconfigurationv1alpha1.AIGatewayConsumer:
 		return AIGatewayConsumerReconciliationWatchOptions(cl)
 	case *aiconfigurationv1alpha1.AIGatewayConsumerCredential: