+++ modules/manager/zz_generated_konnect_index_options.go
@@ -16,6 +16,8 @@ func generatedIndexOptionsForKonnectEntities(
 	return slices.Concat(
 		index.OptionsForAIGatewayAgent(),
 		index.OptionsForAIGatewayAuthStrategy(),
+		index.OptionsForAIGatewayCACertificate(),
+		index.OptionsForAIGatewayCertificate(),
 		index.OptionsForAIGatewayConsumer(),
 		index.OptionsForAIGatewayConsumerCredential(),
 		index.OptionsForAIGatewayConsumerGroup(),