+++ test/mocks/sdkmocks/zz_generated_sdkfactory_mock.go
@@ -12,6 +12,8 @@ import (
 type generatedMockSDKWrapper struct {
 	AIGatewayAgentsSDK                           *mocks.MockAIGatewayAgentsSDK
 	AIGatewayAuthStrategiesSDK                   *mocks.MockAIGatewayAuthStrategiesSDK
+	AIGatewayCACertificatesSDK                   *mocks.MockAIGatewayCACertificatesSDK
+	AIGatewayCertificatesSDK                     *mocks.MockAIGatewayCertificatesSDK
 	AIGatewayConsumersSDK                        *mocks.MockAIGatewayConsumersSDK
 	AIGatewayConsumerGroupsSDK                   *mocks.MockAIGatewayConsumerGroupsSDK
 	AIGatewayDataPlaneCertificatesSDK            *mocks.MockAIGatewayDataPlaneCertificatesSDK
@@ -45,6 +47,8 @@ func newGeneratedMockSDKWrapper(t *testing.T) generatedMockSDKWrapper {
 	return generatedMockSDKWrapper{
 		AIGatewayAgentsSDK:                           mocks.NewMockAIGatewayAgentsSDK(t),
 		AIGatewayAuthStrategiesSDK:                   mocks.NewMockAIGatewayAuthStrategiesSDK(t),
+		AIGatewayCACertificatesSDK:                   mocks.NewMockAIGatewayCACertificatesSDK(t),
+		AIGatewayCertificatesSDK:                     mocks.NewMockAIGatewayCertificatesSDK(t),
 		AIGatewayConsumersSDK:                        mocks.NewMockAIGatewayConsumersSDK(t),
 		AIGatewayConsumerGroupsSDK:                   mocks.NewMockAIGatewayConsumerGroupsSDK(t),
 		AIGatewayDataPlaneCertificatesSDK:            mocks.NewMockAIGatewayDataPlaneCertificatesSDK(t),
@@ -83,6 +87,16 @@ func (m generatedMockSDKWrapper) GetAIGatewayAgentsSDK() sdkkonnectgo.AIGatewayA
 // GetAIGatewayAuthStrategiesSDK returns the SDK to operate AIGatewayAuthStrategy.
 func (m generatedMockSDKWrapper) GetAIGatewayAuthStrategiesSDK() sdkkonnectgo.AIGatewayAuthStrategiesSDK {
 	return m.AIGatewayAuthStrategiesSDK
+}
+
+// GetAIGatewayCACertificatesSDK returns the SDK to operate AIGatewayCACertificate.
+func (m generatedMockSDKWrapper) GetAIGatewayCACertificatesSDK() sdkkonnectgo.AIGatewayCACertificatesSDK {
+	return m.AIGatewayCACertificatesSDK
+}
+
+// GetAIGatewayCertificatesSDK returns the SDK to operate AIGatewayCertificate.
+func (m generatedMockSDKWrapper) GetAIGatewayCertificatesSDK() sdkkonnectgo.AIGatewayCertificatesSDK {
+	return m.AIGatewayCertificatesSDK
 }
 
 // GetAIGatewayConsumersSDK returns the SDK to operate AIGatewayConsumer.